Feature: Add a workaround option for collapsable GIFs
Mattermost doesn't provide a minimize/collapse button for embedded
images. For users that prefer a raw URL, which image previews *has* a
collapse button, we add an option.
Ideally this option will disappear in the future if/when Mattermost
allows to collapse embedded images.

See #12 for the dicussion and MM issue references