Feature: avoid URL duplicate in markdown to avoid duplicate preview
In future releases of Mattermost server (> 5.30.0), makdown links will
also trigger the display of the URL preview, as per
https://mattermost.atlassian.net/browse/MM-29224 .

This means a double preview image is displayed if there's both the
markdown link on the keywords and an embedded image.

This Mattermost evolution will not affect the native client at first so
we have to keep the embedded image and get rid of the URL link on the
keywords. A side effect is that if the image proxy is enabled on the
Mattermost server, users won't be able to access the original URL and
share a link to the GIF that will work outside the Mattermost instance.

It will be interesting to check back on this later, as plugins might get
a configuration option to choos how to display previews.