Drop ARM support to get install package size < 50M
With more than 4 executables, the package is > to the maximum
authorized HTTP body allowed.

As FreeBSD support was requested, but not ARM support (which are very
recent and were not previously supported by this plugin), we drop ARM support
while waiting for a solution by the MM team for this issue (ARM support
was introduced by the starter-template).