Implement pseudo-random GIFs for Tenor and Gfycat
Now, in all cases except Giphy with random option activated, load a page of Gif
results which are added to the shuffle context: if there are still Gifs after the current
index, shuffles uses the next GIF, else it load a new page of results.

Tenor API can randomize the result, but for Gfycat, we need to randomize
the page items ourselves.