Show attribution message only in preview messages (if there is one)
As suggested in #60, if the plugin is configured to show a preview post, which
contains the GIF provider attribution message, then we don't show the
attribution message for the posted GIF message.

If the plugin is configured to post a GIF without the preview message, then
as previously, the attribution message is still shown is the posted GIF
message.