DO NOT MERGE ON MAIN debug CI without cached Go packages
Trying to debug why the testify upgrade breaks the compilation with

"build: server/commands.go#L25
p.API undefined (type *Plugin has no field or method API) (typecheck)"

errors. One idea is that golangci-lint might be unable to retrieve the
typings from the Mattermost repository since Mattermost migrated to a
monorepo (according to this discussion
https://github.com/golangci/golangci-lint/discussions/2287#discussioncomment-1982414
).

Currently, the build works on the main branch. Let's disable the cache
and see if it still works!